Removed unnecessary casts and imports.
git-svn-id: https://svn.apache.org/repos/asf/james/mime4j/trunk@661706 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/james/mime4j/decoder/QuotedPrintableInputStream.java
@@ -121,7 +121,7 @@ public class QuotedPrintableInputStream extends InputStream {
                     return;
             }
 
-            byte b = (byte)pushbackq.dequeue();
+            byte b = pushbackq.dequeue();
 
             switch (state) {
                 case 0:  // start state, no bytes pending
